+++ space_filling_pytorch/functional/z_order.py
@@ -57,7 +57,7 @@ def _encode_z_kernel(
 
     # assign batch index
     if ASSIGN_BATCH_INDEX:
-        ret |= pid_b << 48
+        ret |= pid_b.to(tl.int64) << 48
 
     tl.store(distance_ptr + pid_b * N + offs_n, ret, mask=mask_n)
 
@@ -128,7 +128,7 @@ def _encode_z_unpadded_kernel(
 
     # assign batch index
     if ASSIGN_BATCH_INDEX:
-        ret |= pid << 48
+        ret |= pid.to(tl.int64) << 48
 
     tl.store(distance_ptr + offs_n, ret, mask=mask)
